chore: unify SD card CSV/JSON text-parser stream and file resolution
The CSV and JSON SD card log parsers each carried an identical copy of
the stream/file resolution logic (including the fallback that buffers
a forward-only stream up front since it can't be re-read) and an
identical private EmptySamples() iterator. Any future fix to that
logic would need to be applied twice, and could silently drift between
the two formats.

Extracted the shared logic into a new internal SdCardTextFileSource
helper and had both parsers delegate to it. No public API changed;
each format's own BuildSessionAsync (which is genuinely different)
was left untouched.

Verified: dotnet build 0 warnings/0 errors; full test suite green on
net9.0 and net10.0 (3726/3726 passed, 2 skipped). Test-only-adjacent
internal refactor, no bench validation needed.

Not merging — for review.
